Introduce RolloutResult
