+++ src/benchflow/__init__.py
@@ -46,7 +46,7 @@ from benchflow.environments import (
 from benchflow.job import Job, JobConfig, JobResult, RetryConfig
 from benchflow.metrics import BenchmarkMetrics, collect_metrics
 from benchflow.models import AgentInstallError, AgentTimeoutError, RunResult
-from benchflow.rollouts import Role, RolloutConfig, Scene, Turn
+from benchflow.rollouts import Role, RolloutConfig, RolloutResult, Scene, Turn
 from benchflow.runtime import (
     Agent,
     Environment,
@@ -101,6 +101,7 @@ __all__ = [
     "AgentTimeoutError",
     "RunResult",
     "RolloutConfig",
+    "RolloutResult",
     # Runtime (0.3 primary API)
     "Agent",
     "Environment",
