+++ tests/test_reexport.py
@@ -58,12 +58,14 @@ def test_public_api_reexports():
         AgentInstallError,
         AgentTimeoutError,
         RolloutConfig,
+        RolloutResult,
         RunResult,
         stage_dockerfile_deps,
     )
 
     assert callable(SDK)
     assert callable(RolloutConfig)
+    assert callable(RolloutResult)
     assert callable(RunResult)
     assert callable(AgentInstallError)
     assert callable(AgentTimeoutError)
